Reconstruction project \n Remove java-7
