+++ .gitignore
@@ -21,10 +21,7 @@
 # virtual machine crash logs, see http://www.java.com/en/download/help/error_hotspot.xml
 hs_err_pid*
 .idea/
-src/test/
 out/
 build/
-java-7/src/test/
-java-8/src/test/
 .gradle/*
 *.iml
