+++ build.gradle
@@ -1,13 +1,17 @@
 
+apply plugin: 'java-library'
+
+repositories {
+    jcenter()
+    mavenCentral()
+}
+
+version = '1.0.6'
 
 
-
-allprojects {
-
-    version = '1.0.6'
-
-    repositories {
-        jcenter()
-        mavenCentral()
-    }
+dependencies{
+    // https://mvnrepository.com/artifact/junit/junit
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+    // https://mvnrepository.com/artifact/org.mockito/mockito-core
+    testCompile group: 'org.mockito', name: 'mockito-core', version: '2.1.0'
 }
