+++ settings.gradle
@@ -1,2 +1,1 @@
-rootProject.name  = 'server-client'
-include 'java-8' ,'java-7'
+
