+++ src/main/java/ge/gnio/annotation/Obtain.java
