+++ src/main/java/ge/gnio/exception/IncompatiblePacketException.java
